Add support to allow macros to be automatically registered by plugins that don't have access to the DRootSpy pointer.
For macros that come in a TMemFile with no other objects, cd into 
the sum hists dir before running it so it can use the summed histograms
there.

Automatically detect what histograms are displayed in a canvas after a
macro was run and request those to be updated so the display will update
properly.

Allow user to specify the histograms a macro needs via special comments
so when the macro is run, the requests can be made.


git-svn-id: https://halldsvn.jlab.org/repos/trunk/online/packages/RootSpy@15321 e333a0bb-b310-0410-bd06-bbc7ae8e47d2